% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{addVectorGridTilesDev}
\alias{addVectorGridTilesDev}
\title{Add VectorGrid tiles to map}
\usage{
addVectorGridTilesDev(map)
}
\arguments{
\item{map}{A leaflet htmlwidget}
}
\description{
Add VectorGrid tiles to map
}
