% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{append_colour}
\alias{append_colour}
\title{Append colour code to x.}
\usage{
append_colour(
  x,
  key = c("ZONE", "SUBZONE", "VARIANT"),
  colour = subzones_colours_ref,
  name = "fillColor"
)
}
\arguments{
\item{x}{A data.frame like object.}

\item{key}{A character vector of columns in x to paste together
to create a merge key. Default to \code{c("ZONE", "SUBZONE", "VARIANT")}.}

\item{colour}{A data.table with a predefined key.
Default to \code{subzones_colours_ref} in this package namespace.}

\item{name}{A character string. Name of the column to create
or replace in x. Default to \code{fillColor}.}
}
\value{
x with \code{name} column appended.
}
\description{
Append colour code to x.
}
\details{
Internal dataset \code{zones_colours_ref} and \code{subzones_colours_ref}
use \code{classification} as \code{key}. This classification is either equal
to \code{ZONE} or a combination of \code{ZONE}, \code{SUBZONE} and \code{VARIANT}.
\code{NA} values are replaced by \code{""}.
}
