% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{bccciss_map}
\alias{bccciss_map}
\title{Main CCISS APP leaflet map}
\usage{
bccciss_map(points, xName = "Long", yName = "Lat", render = c("app", "report"))
}
\arguments{
\item{pts}{A data.frame like object containing geoposition of
points of interest in CRS 4326.}
}
\value{
A Leaflet map object.
}
\description{
Main CCISS APP leaflet map
}
