% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{cciss-data}
\alias{cciss-data}
\alias{E1}
\alias{F1}
\alias{R1}
\alias{S1}
\alias{subzones_colours_ref}
\alias{zones_colours_ref}
\title{Data to be included in cciss package}
\description{
Data to be included in cciss package
}
\keyword{data}
