% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility.R
\name{ccissOutput}
\alias{ccissOutput}
\title{ccissOutput}
\usage{
ccissOutput(SSPred, suit, rules, feasFlag, histWeights, futureWeights)
}
\arguments{
\item{SSPred}{A data.table. Predictions input.}

\item{suit}{A data.table. Suitability.}

\item{rules}{A data.table. Set of rules.}

\item{feasFlag}{A data.table. Flag depending on feasibility differential.}

\item{histWeights}{A numeric vector of length 3.}

\item{futureWeights}{A numeric vector of length 4.}
}
\value{
What the function returns
}
\description{
ccissOutput
}
\details{
What the function does
}
