% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_fns.R
\name{cleanData}
\alias{cleanData}
\title{Clean Portfolio Data}
\usage{
cleanData(SSPredAll, SIBEC, SuitTable, SNum, Trees, timePer, selectBGC)
}
\arguments{
\item{SSPredAll}{cleanData parameter.}

\item{SIBEC}{cleanData parameter.}

\item{SuitTable}{cleanData parameter.}

\item{SNum}{cleanData parameter.}

\item{Trees}{cleanData parameter.}

\item{timePer}{cleanData parameter.}

\item{selectBGC}{cleanData parameter.}
}
\value{
What the function returns
}
\description{
Clean Portfolio Data
}
\details{
What the function does
}
