% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_auth.R
\name{db_auth}
\alias{db_auth}
\alias{init_con}
\alias{db_deauth}
\alias{db_auth_check}
\title{Auth methods}
\usage{
init_con(
  host = "localhost",
  dbname = "cciss_data",
  port = 5432,
  user = "postgres",
  ...
)

db_auth(host, dbname, port, user, force = FALSE, ...)

db_deauth()

db_auth_check(host, dbname, port, user, password, ...)
}
\arguments{
\item{host}{A character string. Postgres database host. Default to \code{localhost}.}

\item{dbname}{A character string. Postgres database name. Default to \code{cciss_data}.}

\item{port}{A numeric. Postgres databse port. Default to 5432.}

\item{user}{A character string. Postgres database user. Default to \code{postgres}.}

\item{...}{Additional parameters for dbConnect.}

\item{force}{A boolean. Force reauthentification.}

\item{password}{Password}
}
\value{
A connection pointer to Postgres database.
}
\description{
Authentification management.
}
\details{
Manage authentification credentials using package \code{keyring}

Deauthorize credentials. Remove from memory.
}
