% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbBbox}
\alias{dbBbox}
\title{Buffered bounding box}
\usage{
dbBbox(con, points, buffer)
}
\arguments{
\item{con}{An active postgres DBI connection.}

\item{points}{A data.frame with two columns representing longitude
(Long) and latitude (Lat) with GPS values (EPSG:4326)}

\item{buffer}{A numeric. Buffer distance in km.}
}
\value{
a bounding box 1km around points
}
\description{
Buffered bounding box
}
