% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetBecInfo}
\alias{dbGetBecInfo}
\title{Pull bec_info from lat long inputs}
\usage{
dbGetBecInfo(con, points)
}
\arguments{
\item{con}{An active postgre DBI connection.}

\item{points}{A data.frame with two columns representing longitude and latitude
GPS values (EPSG:4326)}
}
\value{
BEC Information about features
}
\description{
Pull bec_info from lat long inputs
}
