% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetBGCPred}
\alias{dbGetBGCPred}
\title{Return raw predictions for given sitenos}
\usage{
dbGetBGCPred(
  con,
  siteno,
  scenario = c("ssp126", "ssp245", "ssp370"),
  period = c("2021", "2041", "2061", "2081")
)
}
\arguments{
\item{con}{An active postgres DBI connection.}

\item{siteno}{A vector of siteno.}

\item{scenario}{Vector of requested SSPs}

\item{period}{Character vector of requested future periods.}
}
\value{
A data.table containing BGC predictions.
}
\description{
Return raw predictions for given sitenos
}
\details{
Return data from CCISS database for provided SiteNo.
}
