% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetCCISSRaw}
\alias{dbGetCCISSRaw}
\title{Pull individual predictions by district}
\usage{
dbGetCCISSRaw(con, district, gcm, scenario, period)
}
\arguments{
\item{con}{An active postgres DBI connection.}

\item{siteno}{A character vector of siteno.}

\item{avg}{A boolean.}

\item{modWeights}{A data table of gcm and rcp weights.}
}
\value{
A data.table containing CCISS information for each provided SiteNo.
}
\description{
Pull individual predictions by district
}
\details{
Get CCISS for provided SiteNo.
}
