% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_fns.R
\name{dbGetClimSum}
\alias{dbGetClimSum}
\title{Get Climate Summary Data}
\usage{
dbGetClimSum(con, BGC, Scn)
}
\arguments{
\item{con}{A DBI connection object.}

\item{BGC}{A character vector of BGC.}

\item{Scn}{A character vector of length 1 matching a scenario.}
}
\description{
Get Climate Summary Data
}
