% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetHexID}
\alias{dbGetHexID}
\title{Pull HexIDs (SiteNo) from lat long inputs}
\usage{
dbGetHexID(con, points)
}
\arguments{
\item{con}{An active postgre DBI connection.}

\item{points}{A data.frame with two columns representing longitude and latitude
GPS values (EPSG:4326)}
}
\value{
A vector of hex IDs,

BEC Information about features
}
\description{
Pull HexIDs (SiteNo) from lat long inputs
}
\details{
Get hex IDs for provided geolocations.
}
