% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_fns.R
\name{dbGetSppLimits}
\alias{dbGetSppLimits}
\title{Species Limits}
\usage{
dbGetSppLimits(con, SuitTable, Trees)
}
\arguments{
\item{con}{A DBI connection object.}

\item{SuitTable}{A suitability data.frame.}

\item{Trees}{A list of tree species.}
}
\description{
Species Limits
}
