% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbPointInfo}
\alias{dbPointInfo}
\title{Pull points info from lat long inputs}
\usage{
dbPointInfo(con, points)
}
\arguments{
\item{con}{An active postgres DBI connection.}

\item{points}{A data.frame with two columns representing longitude
(Long) and latitude (Lat) with GPS values (EPSG:4326)}
}
\value{
BEC Information, Site Reference Hex Grid Number and elevation in meter
}
\description{
Pull points info from lat long inputs
}
