% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital_ocean.R
\name{setup_docklet}
\alias{setup_docklet}
\alias{launch_tileserver}
\title{Setup droplet}
\usage{
setup_docklet(droplet)

launch_tileserver(droplet)
}
\arguments{
\item{droplet}{DO droplet}
}
\description{
Setup droplet

Launch tileserver GL
}
