% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_fns.R
\name{edatopicSubset}
\alias{edatopicSubset}
\title{Portfolio Subset}
\usage{
edatopicSubset(SSPredOrig, eda, pos = "Zonal")
}
\arguments{
\item{SSPredOrig}{A parameter}

\item{eda}{A parameter}

\item{pos}{A parameter}
}
\description{
Portfolio Subset
}
