% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_fns.R
\name{ef_plot}
\alias{ef_plot}
\title{Plot Efficient Frontier}
\usage{
ef_plot(efAll, intDat, colScale)
}
\arguments{
\item{efAll}{A parameter}

\item{intDat}{A parameter}

\item{colScale}{A parameter}
}
\description{
Plot Efficient Frontier
}
