% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{localGetBecInfo}
\alias{localGetBecInfo}
\title{Pull bec_info from lat long inputs}
\usage{
localGetBecInfo(points, xName = "Long", yName = "Lat", host, ...)
}
\arguments{
\item{points}{A data.frame with columns xName yName representing longitude and latitude
GPS values (EPSG:4326)}

\item{xName}{A character string. Longitude column name.}

\item{yName}{A character string. Latitude column name.}

\item{host}{A character string. Postgres database host. Default to \code{localhost}.}

\item{...}{Additional parameters for dbConnect.}
}
\value{
BEC Information about features
}
\description{
Pull bec_info from lat long inputs
}
