% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ModelDir}
\alias{ModelDir}
\title{Function for quickly calculating model direction/agreement}
\usage{
ModelDir(x, Curr, dir)
}
\arguments{
\item{x}{data}

\item{dir}{Improve or Decline}

\item{curr}{vector of current suitability}
}
\value{
NumericVector
}
\description{
Function for quickly calculating model direction/agreement
}
