% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortfolioOpt_R.R
\name{optimise_portfolio}
\alias{optimise_portfolio}
\title{Optimise Portfolio}
\usage{
optimise_portfolio(returns, cov_matrix, boundDat, minTot)
}
\arguments{
\item{returns}{A data.table of tree growth}

\item{cov_matrix}{A correlation matrix of trees}

\item{boundDat}{A data.table of bounds for each species}

\item{minTot}{A double; minimum acceptable weight}
}
\value{
A data.table with optimised weights across the frontier, the variance, and the return
}
\description{
Optimise Portfolio
}
\details{
The function uses the nloptr library to optimise variance over the efficient frontier
}
\author{
Kiri Daust
}
