% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{plugins_loaded}
\alias{plugins_loaded}
\title{Load extra plugins dependencies on package load}
\usage{
plugins_loaded()
}
\description{
Load extra plugins dependencies on package load
}
\details{
\href{https://github.com/Leaflet/Leaflet.VectorGrid}{VectorGrid}
is a leaflet extension to display gridded vector data
(sliced GeoJSON, TopoJSON or protobuf vector tiles) in Leaflet 1.0.0
}
