% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{registerPlugin}
\alias{registerPlugin}
\title{Additional leaflet plugins registration}
\usage{
registerPlugin(map, plugin)
}
\arguments{
\item{map}{An htmlwidget. i.e. a leaflet map.}

\item{plugin}{An htmlDependency object from htmltools.}
}
\description{
Additional leaflet plugins registration
}
\details{
Adds plugin to the widget.
}
