% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_fns.R
\name{run_portfolio}
\alias{run_portfolio}
\title{Run species Portfolio}
\usage{
run_portfolio(
  SiteList,
  climVar,
  SSPredAll,
  SIBEC,
  SuitTable,
  Trees,
  TimePeriods,
  selectBGC,
  SuitProb,
  returnValue,
  sppLimits,
  minAccept,
  boundDat,
  ProbPest
)
}
\arguments{
\item{SiteList}{A parameter}

\item{climVar}{A parameter}

\item{SSPredAll}{A parameter}

\item{SIBEC}{A parameter}

\item{SuitTable}{A parameter}

\item{Trees}{A parameter}

\item{TimePeriods}{A parameter}

\item{selectBGC}{A parameter}

\item{SuitProb}{A parameter}

\item{returnValue}{A parameter}

\item{sppLimits}{A parameter}

\item{minAccept}{A parameter}

\item{boundDat}{A parameter}

\item{ProbPest}{A parameter}
}
\description{
Run species Portfolio
}
