% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{vector_tiles_provider}
\alias{vector_tiles_provider}
\title{Define VectorGrid tiles provider to use with
VectorGrid.Protobuf}
\usage{
vector_tiles_provider(name, uri_template)
}
\arguments{
\item{name}{A character string. Name of the tiles provider.}

\item{uri_template}{A character string. URI template to fetch tiles.}
}
\value{
A list of provider attributes.
}
\description{
Define VectorGrid tiles provider to use with
VectorGrid.Protobuf
}
\details{
See leaflet
\href{https://leaflet.github.io/Leaflet.VectorGrid/vectorgrid-api-docs.html#vectorgrid-protobuf}{VectorGrid documentation}
for more details
}
