% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{wms_loaded}
\alias{wms_loaded}
\title{Load extra WMS tiles providers on package load}
\usage{
wms_loaded()
}
\description{
Load extra WMS tiles providers on package load
}
\details{
USGS Hill Shade provides simple black and white
relief tiles.
}
