% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{bccciss_map}
\alias{bccciss_map}
\title{Main CCISS APP leaflet map}
\usage{
bccciss_map(points, bec_tiles)
}
\arguments{
\item{points}{A data.frame like object containing geoposition of
points of interest.}

\item{bec_tiles}{A vector tiles provider created with
\link{vector_tiles_provider} to fetch BEC layers.}
}
\value{
A Leaflet map object.
}
\description{
Main CCISS APP leaflet map
}
