% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{bccciss_map}
\alias{bccciss_map}
\title{Main CCISS APP leaflet map}
\usage{
bccciss_map(points, xName = "Long", yName = "Lat", render = c("app", "report"))
}
\arguments{
\item{points}{A data.frame with columns xName yName representing longitude and latitude
GPS values (EPSG:4326)}

\item{xName}{A character string. Longitude column name.}

\item{yName}{A character string. Latitude column name.}

\item{render}{A character. Where the map will be rendered. Either in a Shiny app or a report.
Default to app.}
}
\value{
A Leaflet map object.
}
\description{
Main CCISS APP leaflet map
}
