% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_fns.R
\name{cleanData}
\alias{cleanData}
\title{Clean Portfolio Data}
\usage{
cleanData(SSPredAll, SIBEC, SuitTable, SNum, Trees, timePer, selectBGC)
}
\value{
What the function returns
}
\description{
Clean Portfolio Data
}
\details{
What the function does
}
