% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetCCISS}
\alias{dbGetCCISS}
\title{Pull CCISS from a vector of SiteNo}
\usage{
dbGetCCISS(SiteNo, avg, scn = c("rcp45", "rcp85"), host = "localhost", ...)
}
\arguments{
\item{SiteNo}{A numeric vector of SiteNo.}

\item{avg}{A boolean.}

\item{scn}{A character string. Scenario name. Either \code{rcp45} or \code{rcp85}.}

\item{host}{A character string. Postgres database host. Default to \code{localhost}.}

\item{...}{Additional parameters for dbConnect.}
}
\value{
A data.table containing CCISS information for each provided SiteNo.
}
\description{
Pull CCISS from a vector of SiteNo
}
\details{
Get CCISS for provided SiteNo.
}
