% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetCCISS}
\alias{dbGetCCISS}
\title{Pull CCISS from a vector of SiteNo}
\usage{
dbGetCCISS(con, siteno, avg, scn = c("rcp45", "rcp85"))
}
\arguments{
\item{con}{An active postgres DBI connection.}

\item{siteno}{A character vector of siteno.}

\item{avg}{A boolean.}

\item{scn}{A character string. Scenario name. Either \code{rcp45} or \code{rcp85}.}
}
\value{
A data.table containing CCISS information for each provided SiteNo.
}
\description{
Pull CCISS from a vector of SiteNo
}
\details{
Get CCISS for provided SiteNo.
}
