% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital_ocean.R
\name{setup_docklet}
\alias{setup_docklet}
\alias{remote_shp_tiles}
\alias{tippecanoe_usage}
\alias{tileserver_doc}
\alias{styling_doc}
\alias{reset_ssh_sessions}
\alias{launch_tileserver}
\title{Setup droplet}
\usage{
setup_docklet(
  size = "s-2vcpu-4gb-intel",
  region = "tor1",
  image = "docker-20-04",
  tags = c("bccciss", "tileserver"),
  checks = FALSE,
  ...
)

remote_shp_tiles(
  droplet,
  ...,
  source_dir,
  remote_dir = "/tmp/shp",
  skip_upload = FALSE
)

tippecanoe_usage()

tileserver_doc()

styling_doc()

reset_ssh_sessions()

launch_tileserver(droplet, config, styles)
}
\arguments{
\item{size}{A character string. Size slug identifier. See \code{\link[analogsea:sizes]{analogsea::sizes()}} for a complete
list. Default to \verb{s-2vcpu-4gb-intel}, large enough to serve layers covering B.C. with fast
memory and disk.}

\item{region}{A character string. The unique slug identifier for the region that you wish to
deploy in. See \code{\link[analogsea:regions]{analogsea::regions()}} for a complete list. Default to \code{tor1} in Canada.}

\item{image}{A character string. The image ID of a public or private image, or the unique slug
identifier for a public image. This image should have Docker already installer.
See \code{\link[analogsea:images]{analogsea::images()}} for a complete list. Default to \code{docker-20-04}.}

\item{tags}{A vector of tag names to apply to the Droplet after it is created.
Tag names can either be existing or new tags. Default to \code{c("bccciss", "tileserver")}}

\item{checks}{A boolean. Check that region is \code{tor1} and image is official \code{Docker}. Default
to false.}

\item{...}{additional command line arguments for \code{tippecanoe}}

\item{droplet}{DO droplet}

\item{source_dir}{Directory containing shape files}

\item{remote_dir}{Directory where to upload files on remote. Default to \verb{/tmp/shp}.}

\item{skip_upload}{A boolean. Skip upload and convert to geojson part. Default FALSE.}

\item{config}{Local path to a tileserver config file}

\item{styles}{Local path to Mapbox GL compatible styles (file or dir)}
}
\value{
A droplet.
}
\description{
Setup droplet

Transform shape files to mbtiles

Open browser to tippecanoe usage manual

Open browser to tileserver GL documentation

Open browser to GL styling documentation

Reset SSH sessions

Launch tileserver GL
}
\details{
Setup a droplet with necessary tooling to convert shape files to vector tiles and
host them using tileserver gl.

Look for shape files in \code{source_dir}. Upload \code{source_dir} content to droplet
\code{remote_dir}. Transform shape files to geojson and call \code{tippecanoe} with each layers found
using \code{...} for parametrization creating \code{tiles.mbtiles}. See \code{\link[=tippecanoe_usage]{tippecanoe_usage()}} for
\code{tippecanoe} usage documentation.

Launch a tileserver GL docker image. Also stops
any already running instances, delete server cache and reload nginx.

If config file is provided, it will be uploaded to server and used
to launch tileserver. See https://tileserver.readthedocs.io/en/latest/config.html
for details about config
}
