% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_overlap.R
\name{edatopicOverlap}
\alias{edatopicOverlap}
\title{EDA Topic Overlap}
\usage{
edatopicOverlap(BGC, E1, E1_Phase)
}
\arguments{
\item{BGC}{BGC}

\item{E1}{A data.table?}

\item{E1_Phase}{A phase?}
}
\value{
What the function returns
}
\description{
EDA Topic Overlap
}
\details{
What the function does
}
